Remove pointless `AbstractClient` indirection
Since the removal of the transport client `o.e.c.i.Client` has exactly
one implementation, `o.e.c.i.s.AbstractClient`, which essentially just
defines how each method maps onto a transport action. Likewise for
`AdminClient`, `ClusterAdminClient` and `IndicesAdminClient`. This
commit inlines these implementations into the parent interfaces.
