+++ server/src/main/java/org/elasticsearch/client/internal/node/NodeClient.java
@@ -15,7 +15,6 @@ import org.elasticsearch.action.ActionType;
 import org.elasticsearch.action.support.TransportAction;
 import org.elasticsearch.client.internal.Client;
 import org.elasticsearch.client.internal.RemoteClusterClient;
-import org.elasticsearch.client.internal.support.AbstractClient;
 import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.tasks.Task;
@@ -33,7 +32,7 @@ import java.util.function.Supplier;
 /**
  * Client that executes actions on the local node.
  */
-public class NodeClient extends AbstractClient {
+public class NodeClient extends Client {
 
     private Map<ActionType<? extends ActionResponse>, TransportAction<? extends ActionRequest, ? extends ActionResponse>> actions;
