--- server/src/main/java/org/elasticsearch/client/internal/support/AbstractClient.java
@@ -1,1538 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-package org.elasticsearch.client.internal.support;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.elasticsearch.action.ActionFuture;
-import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.ActionRequest;
-import org.elasticsearch.action.ActionResponse;
-import org.elasticsearch.action.ActionType;
-import org.elasticsearch.action.DocWriteResponse;
-import org.elasticsearch.action.admin.cluster.allocation.ClusterAllocationExplainRequest;
-import org.elasticsearch.action.admin.cluster.allocation.ClusterAllocationExplainRequestBuilder;
-import org.elasticsearch.action.admin.cluster.allocation.ClusterAllocationExplainResponse;
-import org.elasticsearch.action.admin.cluster.allocation.TransportClusterAllocationExplainAction;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthRequest;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthRequestBuilder;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthResponse;
-import org.elasticsearch.action.admin.cluster.health.TransportClusterHealthAction;
-import org.elasticsearch.action.admin.cluster.node.info.NodesInfoRequest;
-import org.elasticsearch.action.admin.cluster.node.info.NodesInfoRequestBuilder;
-import org.elasticsearch.action.admin.cluster.node.info.NodesInfoResponse;
-import org.elasticsearch.action.admin.cluster.node.info.TransportNodesInfoAction;
-import org.elasticsearch.action.admin.cluster.node.stats.NodesStatsRequest;
-import org.elasticsearch.action.admin.cluster.node.stats.NodesStatsRequestBuilder;
-import org.elasticsearch.action.admin.cluster.node.stats.NodesStatsResponse;
-import org.elasticsearch.action.admin.cluster.node.stats.TransportNodesStatsAction;
-import org.elasticsearch.action.admin.cluster.node.tasks.cancel.CancelTasksRequest;
-import org.elasticsearch.action.admin.cluster.node.tasks.cancel.CancelTasksRequestBuilder;
-import org.elasticsearch.action.admin.cluster.node.tasks.cancel.TransportCancelTasksAction;
-import org.elasticsearch.action.admin.cluster.node.tasks.get.GetTaskRequest;
-import org.elasticsearch.action.admin.cluster.node.tasks.get.GetTaskRequestBuilder;
-import org.elasticsearch.action.admin.cluster.node.tasks.get.GetTaskResponse;
-import org.elasticsearch.action.admin.cluster.node.tasks.get.TransportGetTaskAction;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksRequest;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksRequestBuilder;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksResponse;
-import org.elasticsearch.action.admin.cluster.node.tasks.list.TransportListTasksAction;
-import org.elasticsearch.action.admin.cluster.node.usage.NodesUsageRequest;
-import org.elasticsearch.action.admin.cluster.node.usage.NodesUsageResponse;
-import org.elasticsearch.action.admin.cluster.node.usage.TransportNodesUsageAction;
-import org.elasticsearch.action.admin.cluster.repositories.cleanup.CleanupRepositoryRequest;
-import org.elasticsearch.action.admin.cluster.repositories.cleanup.CleanupRepositoryRequestBuilder;
-import org.elasticsearch.action.admin.cluster.repositories.cleanup.CleanupRepositoryResponse;
-import org.elasticsearch.action.admin.cluster.repositories.cleanup.TransportCleanupRepositoryAction;
-import org.elasticsearch.action.admin.cluster.repositories.delete.DeleteRepositoryRequest;
-import org.elasticsearch.action.admin.cluster.repositories.delete.DeleteRepositoryRequestBuilder;
-import org.elasticsearch.action.admin.cluster.repositories.delete.TransportDeleteRepositoryAction;
-import org.elasticsearch.action.admin.cluster.repositories.get.GetRepositoriesAction;
-import org.elasticsearch.action.admin.cluster.repositories.get.GetRepositoriesRequest;
-import org.elasticsearch.action.admin.cluster.repositories.get.GetRepositoriesRequestBuilder;
-import org.elasticsearch.action.admin.cluster.repositories.get.GetRepositoriesResponse;
-import org.elasticsearch.action.admin.cluster.repositories.put.PutRepositoryRequest;
-import org.elasticsearch.action.admin.cluster.repositories.put.PutRepositoryRequestBuilder;
-import org.elasticsearch.action.admin.cluster.repositories.put.TransportPutRepositoryAction;
-import org.elasticsearch.action.admin.cluster.repositories.verify.VerifyRepositoryAction;
-import org.elasticsearch.action.admin.cluster.repositories.verify.VerifyRepositoryRequest;
-import org.elasticsearch.action.admin.cluster.repositories.verify.VerifyRepositoryRequestBuilder;
-import org.elasticsearch.action.admin.cluster.repositories.verify.VerifyRepositoryResponse;
-import org.elasticsearch.action.admin.cluster.reroute.ClusterRerouteRequest;
-import org.elasticsearch.action.admin.cluster.reroute.ClusterRerouteRequestBuilder;
-import org.elasticsearch.action.admin.cluster.reroute.ClusterRerouteResponse;
-import org.elasticsearch.action.admin.cluster.reroute.TransportClusterRerouteAction;
-import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsAction;
-import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
-import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequestBuilder;
-import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsResponse;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequestBuilder;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
-import org.elasticsearch.action.admin.cluster.shards.TransportClusterSearchShardsAction;
-import org.elasticsearch.action.admin.cluster.snapshots.clone.CloneSnapshotRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.clone.CloneSnapshotRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.clone.TransportCloneSnapshotAction;
-import org.elasticsearch.action.admin.cluster.snapshots.create.CreateSnapshotRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.create.CreateSnapshotRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.create.CreateSnapshotResponse;
-import org.elasticsearch.action.admin.cluster.snapshots.create.TransportCreateSnapshotAction;
-import org.elasticsearch.action.admin.cluster.snapshots.delete.DeleteSnapshotRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.delete.DeleteSnapshotRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.delete.TransportDeleteSnapshotAction;
-import org.elasticsearch.action.admin.cluster.snapshots.get.GetSnapshotsRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.get.GetSnapshotsRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.get.GetSnapshotsResponse;
-import org.elasticsearch.action.admin.cluster.snapshots.get.TransportGetSnapshotsAction;
-import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotResponse;
-import org.elasticsearch.action.admin.cluster.snapshots.restore.TransportRestoreSnapshotAction;
-import org.elasticsearch.action.admin.cluster.snapshots.status.SnapshotsStatusRequest;
-import org.elasticsearch.action.admin.cluster.snapshots.status.SnapshotsStatusRequestBuilder;
-import org.elasticsearch.action.admin.cluster.snapshots.status.SnapshotsStatusResponse;
-import org.elasticsearch.action.admin.cluster.snapshots.status.TransportSnapshotsStatusAction;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateAction;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateRequest;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateRequestBuilder;
-import org.elasticsearch.action.admin.cluster.state.ClusterStateResponse;
-import org.elasticsearch.action.admin.cluster.stats.ClusterStatsRequest;
-import org.elasticsearch.action.admin.cluster.stats.ClusterStatsRequestBuilder;
-import org.elasticsearch.action.admin.cluster.stats.ClusterStatsResponse;
-import org.elasticsearch.action.admin.cluster.stats.TransportClusterStatsAction;
-import org.elasticsearch.action.admin.cluster.storedscripts.DeleteStoredScriptRequest;
-import org.elasticsearch.action.admin.cluster.storedscripts.DeleteStoredScriptRequestBuilder;
-import org.elasticsearch.action.admin.cluster.storedscripts.GetStoredScriptAction;
-import org.elasticsearch.action.admin.cluster.storedscripts.GetStoredScriptRequest;
-import org.elasticsearch.action.admin.cluster.storedscripts.GetStoredScriptRequestBuilder;
-import org.elasticsearch.action.admin.cluster.storedscripts.GetStoredScriptResponse;
-import org.elasticsearch.action.admin.cluster.storedscripts.PutStoredScriptRequest;
-import org.elasticsearch.action.admin.cluster.storedscripts.PutStoredScriptRequestBuilder;
-import org.elasticsearch.action.admin.cluster.storedscripts.TransportDeleteStoredScriptAction;
-import org.elasticsearch.action.admin.cluster.storedscripts.TransportPutStoredScriptAction;
-import org.elasticsearch.action.admin.indices.alias.IndicesAliasesRequest;
-import org.elasticsearch.action.admin.indices.alias.IndicesAliasesRequestBuilder;
-import org.elasticsearch.action.admin.indices.alias.IndicesAliasesResponse;
-import org.elasticsearch.action.admin.indices.alias.TransportIndicesAliasesAction;
-import org.elasticsearch.action.admin.indices.alias.get.GetAliasesAction;
-import org.elasticsearch.action.admin.indices.alias.get.GetAliasesRequest;
-import org.elasticsearch.action.admin.indices.alias.get.GetAliasesRequestBuilder;
-import org.elasticsearch.action.admin.indices.alias.get.GetAliasesResponse;
-import org.elasticsearch.action.admin.indices.analyze.AnalyzeAction;
-import org.elasticsearch.action.admin.indices.analyze.AnalyzeRequestBuilder;
-import org.elasticsearch.action.admin.indices.cache.clear.ClearIndicesCacheRequest;
-import org.elasticsearch.action.admin.indices.cache.clear.ClearIndicesCacheRequestBuilder;
-import org.elasticsearch.action.admin.indices.cache.clear.TransportClearIndicesCacheAction;
-import org.elasticsearch.action.admin.indices.close.CloseIndexRequest;
-import org.elasticsearch.action.admin.indices.close.CloseIndexRequestBuilder;
-import org.elasticsearch.action.admin.indices.close.CloseIndexResponse;
-import org.elasticsearch.action.admin.indices.close.TransportCloseIndexAction;
-import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
-import org.elasticsearch.action.admin.indices.create.CreateIndexRequestBuilder;
-import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
-import org.elasticsearch.action.admin.indices.create.TransportCreateIndexAction;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequestBuilder;
-import org.elasticsearch.action.admin.indices.delete.TransportDeleteIndexAction;
-import org.elasticsearch.action.admin.indices.flush.FlushAction;
-import org.elasticsearch.action.admin.indices.flush.FlushRequest;
-import org.elasticsearch.action.admin.indices.flush.FlushRequestBuilder;
-import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeAction;
-import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeRequest;
-import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeRequestBuilder;
-import org.elasticsearch.action.admin.indices.get.GetIndexAction;
-import org.elasticsearch.action.admin.indices.get.GetIndexRequest;
-import org.elasticsearch.action.admin.indices.get.GetIndexRequestBuilder;
-import org.elasticsearch.action.admin.indices.get.GetIndexResponse;
-import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsAction;
-import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsRequest;
-import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsRequestBuilder;
-import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsResponse;
-import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsAction;
-import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsRequest;
-import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsRequestBuilder;
-import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsResponse;
-import org.elasticsearch.action.admin.indices.mapping.put.PutMappingRequest;
-import org.elasticsearch.action.admin.indices.mapping.put.PutMappingRequestBuilder;
-import org.elasticsearch.action.admin.indices.mapping.put.TransportPutMappingAction;
-import org.elasticsearch.action.admin.indices.open.OpenIndexAction;
-import org.elasticsearch.action.admin.indices.open.OpenIndexRequest;
-import org.elasticsearch.action.admin.indices.open.OpenIndexRequestBuilder;
-import org.elasticsearch.action.admin.indices.open.OpenIndexResponse;
-import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockRequest;
-import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockRequestBuilder;
-import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockResponse;
-import org.elasticsearch.action.admin.indices.readonly.TransportAddIndexBlockAction;
-import org.elasticsearch.action.admin.indices.recovery.RecoveryAction;
-import org.elasticsearch.action.admin.indices.recovery.RecoveryRequest;
-import org.elasticsearch.action.admin.indices.recovery.RecoveryRequestBuilder;
-import org.elasticsearch.action.admin.indices.recovery.RecoveryResponse;
-import org.elasticsearch.action.admin.indices.refresh.RefreshAction;
-import org.elasticsearch.action.admin.indices.refresh.RefreshRequest;
-import org.elasticsearch.action.admin.indices.refresh.RefreshRequestBuilder;
-import org.elasticsearch.action.admin.indices.resolve.ResolveIndexAction;
-import org.elasticsearch.action.admin.indices.rollover.RolloverAction;
-import org.elasticsearch.action.admin.indices.rollover.RolloverRequest;
-import org.elasticsearch.action.admin.indices.rollover.RolloverRequestBuilder;
-import org.elasticsearch.action.admin.indices.rollover.RolloverResponse;
-import org.elasticsearch.action.admin.indices.segments.IndicesSegmentResponse;
-import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsAction;
-import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsRequest;
-import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsRequestBuilder;
-import org.elasticsearch.action.admin.indices.settings.get.GetSettingsAction;
-import org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequest;
-import org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequestBuilder;
-import org.elasticsearch.action.admin.indices.settings.get.GetSettingsResponse;
-import org.elasticsearch.action.admin.indices.settings.put.TransportUpdateSettingsAction;
-import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsRequest;
-import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsRequestBuilder;
-import org.elasticsearch.action.admin.indices.shrink.ResizeAction;
-import org.elasticsearch.action.admin.indices.shrink.ResizeRequest;
-import org.elasticsearch.action.admin.indices.shrink.ResizeRequestBuilder;
-import org.elasticsearch.action.admin.indices.stats.IndicesStatsAction;
-import org.elasticsearch.action.admin.indices.stats.IndicesStatsRequest;
-import org.elasticsearch.action.admin.indices.stats.IndicesStatsRequestBuilder;
-import org.elasticsearch.action.admin.indices.stats.IndicesStatsResponse;
-import org.elasticsearch.action.admin.indices.template.delete.DeleteIndexTemplateRequest;
-import org.elasticsearch.action.admin.indices.template.delete.DeleteIndexTemplateRequestBuilder;
-import org.elasticsearch.action.admin.indices.template.delete.TransportDeleteIndexTemplateAction;
-import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesAction;
-import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesRequest;
-import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesRequestBuilder;
-import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesResponse;
-import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequest;
-import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequestBuilder;
-import org.elasticsearch.action.admin.indices.template.put.TransportPutIndexTemplateAction;
-import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryAction;
-import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryRequest;
-import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryRequestBuilder;
-import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryResponse;
-import org.elasticsearch.action.bulk.BulkRequest;
-import org.elasticsearch.action.bulk.BulkRequestBuilder;
-import org.elasticsearch.action.bulk.BulkResponse;
-import org.elasticsearch.action.bulk.TransportBulkAction;
-import org.elasticsearch.action.delete.DeleteRequest;
-import org.elasticsearch.action.delete.DeleteRequestBuilder;
-import org.elasticsearch.action.delete.DeleteResponse;
-import org.elasticsearch.action.delete.TransportDeleteAction;
-import org.elasticsearch.action.explain.ExplainRequest;
-import org.elasticsearch.action.explain.ExplainRequestBuilder;
-import org.elasticsearch.action.explain.ExplainResponse;
-import org.elasticsearch.action.explain.TransportExplainAction;
-import org.elasticsearch.action.fieldcaps.FieldCapabilitiesRequest;
-import org.elasticsearch.action.fieldcaps.FieldCapabilitiesRequestBuilder;
-import org.elasticsearch.action.fieldcaps.FieldCapabilitiesResponse;
-import org.elasticsearch.action.fieldcaps.TransportFieldCapabilitiesAction;
-import org.elasticsearch.action.get.GetRequest;
-import org.elasticsearch.action.get.GetRequestBuilder;
-import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.action.get.MultiGetRequest;
-import org.elasticsearch.action.get.MultiGetRequestBuilder;
-import org.elasticsearch.action.get.MultiGetResponse;
-import org.elasticsearch.action.get.TransportGetAction;
-import org.elasticsearch.action.get.TransportMultiGetAction;
-import org.elasticsearch.action.index.IndexRequest;
-import org.elasticsearch.action.index.IndexRequestBuilder;
-import org.elasticsearch.action.index.TransportIndexAction;
-import org.elasticsearch.action.ingest.DeletePipelineRequest;
-import org.elasticsearch.action.ingest.DeletePipelineRequestBuilder;
-import org.elasticsearch.action.ingest.DeletePipelineTransportAction;
-import org.elasticsearch.action.ingest.GetPipelineAction;
-import org.elasticsearch.action.ingest.GetPipelineRequest;
-import org.elasticsearch.action.ingest.GetPipelineRequestBuilder;
-import org.elasticsearch.action.ingest.GetPipelineResponse;
-import org.elasticsearch.action.ingest.PutPipelineRequest;
-import org.elasticsearch.action.ingest.PutPipelineRequestBuilder;
-import org.elasticsearch.action.ingest.PutPipelineTransportAction;
-import org.elasticsearch.action.ingest.SimulatePipelineAction;
-import org.elasticsearch.action.ingest.SimulatePipelineRequest;
-import org.elasticsearch.action.ingest.SimulatePipelineRequestBuilder;
-import org.elasticsearch.action.ingest.SimulatePipelineResponse;
-import org.elasticsearch.action.search.ClearScrollRequest;
-import org.elasticsearch.action.search.ClearScrollRequestBuilder;
-import org.elasticsearch.action.search.ClearScrollResponse;
-import org.elasticsearch.action.search.MultiSearchRequest;
-import org.elasticsearch.action.search.MultiSearchRequestBuilder;
-import org.elasticsearch.action.search.MultiSearchResponse;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.action.search.SearchRequestBuilder;
-import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.search.SearchScrollRequest;
-import org.elasticsearch.action.search.SearchScrollRequestBuilder;
-import org.elasticsearch.action.search.TransportClearScrollAction;
-import org.elasticsearch.action.search.TransportMultiSearchAction;
-import org.elasticsearch.action.search.TransportSearchAction;
-import org.elasticsearch.action.search.TransportSearchScrollAction;
-import org.elasticsearch.action.support.PlainActionFuture;
-import org.elasticsearch.action.support.broadcast.BroadcastResponse;
-import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.action.termvectors.MultiTermVectorsAction;
-import org.elasticsearch.action.termvectors.MultiTermVectorsRequest;
-import org.elasticsearch.action.termvectors.MultiTermVectorsRequestBuilder;
-import org.elasticsearch.action.termvectors.MultiTermVectorsResponse;
-import org.elasticsearch.action.termvectors.TermVectorsAction;
-import org.elasticsearch.action.termvectors.TermVectorsRequest;
-import org.elasticsearch.action.termvectors.TermVectorsRequestBuilder;
-import org.elasticsearch.action.termvectors.TermVectorsResponse;
-import org.elasticsearch.action.update.TransportUpdateAction;
-import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.action.update.UpdateRequestBuilder;
-import org.elasticsearch.action.update.UpdateResponse;
-import org.elasticsearch.client.internal.AdminClient;
-import org.elasticsearch.client.internal.Client;
-import org.elasticsearch.client.internal.ClusterAdminClient;
-import org.elasticsearch.client.internal.ElasticsearchClient;
-import org.elasticsearch.client.internal.FilterClient;
-import org.elasticsearch.client.internal.IndicesAdminClient;
-import org.elasticsearch.cluster.metadata.IndexMetadata.APIBlock;
-import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.util.concurrent.ThreadContext;
-import org.elasticsearch.core.Nullable;
-import org.elasticsearch.core.RefCounted;
-import org.elasticsearch.tasks.TaskId;
-import org.elasticsearch.threadpool.ThreadPool;
-import org.elasticsearch.xcontent.XContentType;
-
-import java.util.Map;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-public abstract class AbstractClient implements Client {
-
-    protected final Logger logger;
-
-    protected final Settings settings;
-    private final ThreadPool threadPool;
-    private final Admin admin;
-
-    public AbstractClient(Settings settings, ThreadPool threadPool) {
-        this.settings = settings;
-        this.threadPool = threadPool;
-        this.admin = new Admin(this);
-        this.logger = LogManager.getLogger(this.getClass());
-    }
-
-    @Override
-    public final Settings settings() {
-        return this.settings;
-    }
-
-    @Override
-    public final ThreadPool threadPool() {
-        return this.threadPool;
-    }
-
-    @Override
-    public final AdminClient admin() {
-        return admin;
-    }
-
-    @Override
-    public final <Request extends ActionRequest, Response extends ActionResponse> ActionFuture<Response> execute(
-        ActionType<Response> action,
-        Request request
-    ) {
-        PlainActionFuture<Response> actionFuture = new RefCountedFuture<>();
-        execute(action, request, actionFuture);
-        return actionFuture;
-    }
-
-    /**
-     * This is the single execution point of *all* clients.
-     */
-    @Override
-    public final <Request extends ActionRequest, Response extends ActionResponse> void execute(
-        ActionType<Response> action,
-        Request request,
-        ActionListener<Response> listener
-    ) {
-        try {
-            doExecute(action, request, listener);
-        } catch (Exception e) {
-            assert false : new AssertionError(e);
-            listener.onFailure(e);
-        }
-    }
-
-    protected abstract <Request extends ActionRequest, Response extends ActionResponse> void doExecute(
-        ActionType<Response> action,
-        Request request,
-        ActionListener<Response> listener
-    );
-
-    @Override
-    public ActionFuture<DocWriteResponse> index(final IndexRequest request) {
-        return execute(TransportIndexAction.TYPE, request);
-    }
-
-    @Override
-    public void index(final IndexRequest request, final ActionListener<DocWriteResponse> listener) {
-        execute(TransportIndexAction.TYPE, request, listener);
-    }
-
-    @Override
-    public IndexRequestBuilder prepareIndex() {
-        return new IndexRequestBuilder(this, null);
-    }
-
-    @Override
-    public IndexRequestBuilder prepareIndex(String index) {
-        return new IndexRequestBuilder(this, index);
-    }
-
-    @Override
-    public ActionFuture<UpdateResponse> update(final UpdateRequest request) {
-        return execute(TransportUpdateAction.TYPE, request);
-    }
-
-    @Override
-    public void update(final UpdateRequest request, final ActionListener<UpdateResponse> listener) {
-        execute(TransportUpdateAction.TYPE, request, listener);
-    }
-
-    @Override
-    public UpdateRequestBuilder prepareUpdate() {
-        return new UpdateRequestBuilder(this, null, null);
-    }
-
-    @Override
-    public UpdateRequestBuilder prepareUpdate(String index, String id) {
-        return new UpdateRequestBuilder(this, index, id);
-    }
-
-    @Override
-    public ActionFuture<DeleteResponse> delete(final DeleteRequest request) {
-        return execute(TransportDeleteAction.TYPE, request);
-    }
-
-    @Override
-    public void delete(final DeleteRequest request, final ActionListener<DeleteResponse> listener) {
-        execute(TransportDeleteAction.TYPE, request, listener);
-    }
-
-    @Override
-    public DeleteRequestBuilder prepareDelete() {
-        return new DeleteRequestBuilder(this, null);
-    }
-
-    @Override
-    public DeleteRequestBuilder prepareDelete(String index, String id) {
-        return prepareDelete().setIndex(index).setId(id);
-    }
-
-    @Override
-    public ActionFuture<BulkResponse> bulk(final BulkRequest request) {
-        return execute(TransportBulkAction.TYPE, request);
-    }
-
-    @Override
-    public void bulk(final BulkRequest request, final ActionListener<BulkResponse> listener) {
-        execute(TransportBulkAction.TYPE, request, listener);
-    }
-
-    @Override
-    public BulkRequestBuilder prepareBulk() {
-        return new BulkRequestBuilder(this);
-    }
-
-    @Override
-    public BulkRequestBuilder prepareBulk(@Nullable String globalIndex) {
-        return new BulkRequestBuilder(this, globalIndex);
-    }
-
-    @Override
-    public ActionFuture<GetResponse> get(final GetRequest request) {
-        return execute(TransportGetAction.TYPE, request);
-    }
-
-    @Override
-    public void get(final GetRequest request, final ActionListener<GetResponse> listener) {
-        execute(TransportGetAction.TYPE, request, listener);
-    }
-
-    @Override
-    public GetRequestBuilder prepareGet() {
-        return new GetRequestBuilder(this, null);
-    }
-
-    @Override
-    public GetRequestBuilder prepareGet(String index, String id) {
-        return prepareGet().setIndex(index).setId(id);
-    }
-
-    @Override
-    public ActionFuture<MultiGetResponse> multiGet(final MultiGetRequest request) {
-        return execute(TransportMultiGetAction.TYPE, request);
-    }
-
-    @Override
-    public void multiGet(final MultiGetRequest request, final ActionListener<MultiGetResponse> listener) {
-        execute(TransportMultiGetAction.TYPE, request, listener);
-    }
-
-    @Override
-    public MultiGetRequestBuilder prepareMultiGet() {
-        return new MultiGetRequestBuilder(this);
-    }
-
-    @Override
-    public ActionFuture<SearchResponse> search(final SearchRequest request) {
-        return execute(TransportSearchAction.TYPE, request);
-    }
-
-    @Override
-    public void search(final SearchRequest request, final ActionListener<SearchResponse> listener) {
-        execute(TransportSearchAction.TYPE, request, listener);
-    }
-
-    @Override
-    public SearchRequestBuilder prepareSearch(String... indices) {
-        return new SearchRequestBuilder(this).setIndices(indices);
-    }
-
-    @Override
-    public ActionFuture<SearchResponse> searchScroll(final SearchScrollRequest request) {
-        return execute(TransportSearchScrollAction.TYPE, request);
-    }
-
-    @Override
-    public void searchScroll(final SearchScrollRequest request, final ActionListener<SearchResponse> listener) {
-        execute(TransportSearchScrollAction.TYPE, request, listener);
-    }
-
-    @Override
-    public SearchScrollRequestBuilder prepareSearchScroll(String scrollId) {
-        return new SearchScrollRequestBuilder(this, scrollId);
-    }
-
-    @Override
-    public ActionFuture<MultiSearchResponse> multiSearch(MultiSearchRequest request) {
-        return execute(TransportMultiSearchAction.TYPE, request);
-    }
-
-    @Override
-    public void multiSearch(MultiSearchRequest request, ActionListener<MultiSearchResponse> listener) {
-        execute(TransportMultiSearchAction.TYPE, request, listener);
-    }
-
-    @Override
-    public MultiSearchRequestBuilder prepareMultiSearch() {
-        return new MultiSearchRequestBuilder(this);
-    }
-
-    @Override
-    public ActionFuture<TermVectorsResponse> termVectors(final TermVectorsRequest request) {
-        return execute(TermVectorsAction.INSTANCE, request);
-    }
-
-    @Override
-    public void termVectors(final TermVectorsRequest request, final ActionListener<TermVectorsResponse> listener) {
-        execute(TermVectorsAction.INSTANCE, request, listener);
-    }
-
-    @Override
-    public TermVectorsRequestBuilder prepareTermVectors() {
-        return new TermVectorsRequestBuilder(this);
-    }
-
-    @Override
-    public TermVectorsRequestBuilder prepareTermVectors(String index, String id) {
-        return new TermVectorsRequestBuilder(this, index, id);
-    }
-
-    @Override
-    public ActionFuture<MultiTermVectorsResponse> multiTermVectors(final MultiTermVectorsRequest request) {
-        return execute(MultiTermVectorsAction.INSTANCE, request);
-    }
-
-    @Override
-    public void multiTermVectors(final MultiTermVectorsRequest request, final ActionListener<MultiTermVectorsResponse> listener) {
-        execute(MultiTermVectorsAction.INSTANCE, request, listener);
-    }
-
-    @Override
-    public MultiTermVectorsRequestBuilder prepareMultiTermVectors() {
-        return new MultiTermVectorsRequestBuilder(this);
-    }
-
-    @Override
-    public ExplainRequestBuilder prepareExplain(String index, String id) {
-        return new ExplainRequestBuilder(this, index, id);
-    }
-
-    @Override
-    public ActionFuture<ExplainResponse> explain(ExplainRequest request) {
-        return execute(TransportExplainAction.TYPE, request);
-    }
-
-    @Override
-    public void explain(ExplainRequest request, ActionListener<ExplainResponse> listener) {
-        execute(TransportExplainAction.TYPE, request, listener);
-    }
-
-    @Override
-    public void clearScroll(ClearScrollRequest request, ActionListener<ClearScrollResponse> listener) {
-        execute(TransportClearScrollAction.TYPE, request, listener);
-    }
-
-    @Override
-    public ActionFuture<ClearScrollResponse> clearScroll(ClearScrollRequest request) {
-        return execute(TransportClearScrollAction.TYPE, request);
-    }
-
-    @Override
-    public ClearScrollRequestBuilder prepareClearScroll() {
-        return new ClearScrollRequestBuilder(this);
-    }
-
-    @Override
-    public void fieldCaps(FieldCapabilitiesRequest request, ActionListener<FieldCapabilitiesResponse> listener) {
-        execute(TransportFieldCapabilitiesAction.TYPE, request, listener);
-    }
-
-    @Override
-    public ActionFuture<FieldCapabilitiesResponse> fieldCaps(FieldCapabilitiesRequest request) {
-        return execute(TransportFieldCapabilitiesAction.TYPE, request);
-    }
-
-    @Override
-    public FieldCapabilitiesRequestBuilder prepareFieldCaps(String... indices) {
-        return new FieldCapabilitiesRequestBuilder(this, indices);
-    }
-
-    static class Admin implements AdminClient {
-
-        private final ClusterAdmin clusterAdmin;
-        private final IndicesAdmin indicesAdmin;
-
-        Admin(ElasticsearchClient client) {
-            this.clusterAdmin = new ClusterAdmin(client);
-            this.indicesAdmin = new IndicesAdmin(client);
-        }
-
-        @Override
-        public ClusterAdminClient cluster() {
-            return clusterAdmin;
-        }
-
-        @Override
-        public IndicesAdminClient indices() {
-            return indicesAdmin;
-        }
-    }
-
-    static class ClusterAdmin implements ClusterAdminClient {
-
-        private final ElasticsearchClient client;
-
-        ClusterAdmin(ElasticsearchClient client) {
-            this.client = client;
-        }
-
-        @Override
-        public <Request extends ActionRequest, Response extends ActionResponse> ActionFuture<Response> execute(
-            ActionType<Response> action,
-            Request request
-        ) {
-            return client.execute(action, request);
-        }
-
-        @Override
-        public <Request extends ActionRequest, Response extends ActionResponse> void execute(
-            ActionType<Response> action,
-            Request request,
-            ActionListener<Response> listener
-        ) {
-            client.execute(action, request, listener);
-        }
-
-        @Override
-        public ThreadPool threadPool() {
-            return client.threadPool();
-        }
-
-        @Override
-        public ActionFuture<ClusterHealthResponse> health(final ClusterHealthRequest request) {
-            return execute(TransportClusterHealthAction.TYPE, request);
-        }
-
-        @Override
-        public void health(final ClusterHealthRequest request, final ActionListener<ClusterHealthResponse> listener) {
-            execute(TransportClusterHealthAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ClusterHealthRequestBuilder prepareHealth(String... indices) {
-            return new ClusterHealthRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<ClusterStateResponse> state(final ClusterStateRequest request) {
-            return execute(ClusterStateAction.INSTANCE, request);
-        }
-
-        @Override
-        public void state(final ClusterStateRequest request, final ActionListener<ClusterStateResponse> listener) {
-            execute(ClusterStateAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ClusterStateRequestBuilder prepareState() {
-            return new ClusterStateRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<ClusterRerouteResponse> reroute(final ClusterRerouteRequest request) {
-            return execute(TransportClusterRerouteAction.TYPE, request);
-        }
-
-        @Override
-        public void reroute(final ClusterRerouteRequest request, final ActionListener<ClusterRerouteResponse> listener) {
-            execute(TransportClusterRerouteAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ClusterRerouteRequestBuilder prepareReroute() {
-            return new ClusterRerouteRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<ClusterUpdateSettingsResponse> updateSettings(final ClusterUpdateSettingsRequest request) {
-            return execute(ClusterUpdateSettingsAction.INSTANCE, request);
-        }
-
-        @Override
-        public void updateSettings(
-            final ClusterUpdateSettingsRequest request,
-            final ActionListener<ClusterUpdateSettingsResponse> listener
-        ) {
-            execute(ClusterUpdateSettingsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ClusterUpdateSettingsRequestBuilder prepareUpdateSettings() {
-            return new ClusterUpdateSettingsRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<NodesInfoResponse> nodesInfo(final NodesInfoRequest request) {
-            return execute(TransportNodesInfoAction.TYPE, request);
-        }
-
-        @Override
-        public void nodesInfo(final NodesInfoRequest request, final ActionListener<NodesInfoResponse> listener) {
-            execute(TransportNodesInfoAction.TYPE, request, listener);
-        }
-
-        @Override
-        public NodesInfoRequestBuilder prepareNodesInfo(String... nodesIds) {
-            return new NodesInfoRequestBuilder(this).setNodesIds(nodesIds);
-        }
-
-        @Override
-        public ActionFuture<NodesStatsResponse> nodesStats(final NodesStatsRequest request) {
-            return execute(TransportNodesStatsAction.TYPE, request);
-        }
-
-        @Override
-        public void nodesStats(final NodesStatsRequest request, final ActionListener<NodesStatsResponse> listener) {
-            execute(TransportNodesStatsAction.TYPE, request, listener);
-        }
-
-        @Override
-        public NodesStatsRequestBuilder prepareNodesStats(String... nodesIds) {
-            return new NodesStatsRequestBuilder(this).setNodesIds(nodesIds);
-        }
-
-        @Override
-        public void nodesUsage(final NodesUsageRequest request, final ActionListener<NodesUsageResponse> listener) {
-            execute(TransportNodesUsageAction.TYPE, request, listener);
-        }
-
-        @Override
-        public void clusterStats(ClusterStatsRequest request, ActionListener<ClusterStatsResponse> listener) {
-            execute(TransportClusterStatsAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ClusterStatsRequestBuilder prepareClusterStats() {
-            return new ClusterStatsRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<ListTasksResponse> listTasks(final ListTasksRequest request) {
-            return execute(TransportListTasksAction.TYPE, request);
-        }
-
-        @Override
-        public void listTasks(final ListTasksRequest request, final ActionListener<ListTasksResponse> listener) {
-            execute(TransportListTasksAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ListTasksRequestBuilder prepareListTasks(String... nodesIds) {
-            return new ListTasksRequestBuilder(this).setNodesIds(nodesIds);
-        }
-
-        @Override
-        public ActionFuture<GetTaskResponse> getTask(final GetTaskRequest request) {
-            return execute(TransportGetTaskAction.TYPE, request);
-        }
-
-        @Override
-        public void getTask(final GetTaskRequest request, final ActionListener<GetTaskResponse> listener) {
-            execute(TransportGetTaskAction.TYPE, request, listener);
-        }
-
-        @Override
-        public GetTaskRequestBuilder prepareGetTask(String taskId) {
-            return prepareGetTask(new TaskId(taskId));
-        }
-
-        @Override
-        public GetTaskRequestBuilder prepareGetTask(TaskId taskId) {
-            return new GetTaskRequestBuilder(this).setTaskId(taskId);
-        }
-
-        @Override
-        public ActionFuture<ListTasksResponse> cancelTasks(CancelTasksRequest request) {
-            return execute(TransportCancelTasksAction.TYPE, request);
-        }
-
-        @Override
-        public void cancelTasks(CancelTasksRequest request, ActionListener<ListTasksResponse> listener) {
-            execute(TransportCancelTasksAction.TYPE, request, listener);
-        }
-
-        @Override
-        public CancelTasksRequestBuilder prepareCancelTasks(String... nodesIds) {
-            return new CancelTasksRequestBuilder(this).setNodesIds(nodesIds);
-        }
-
-        @Override
-        public void searchShards(final ClusterSearchShardsRequest request, final ActionListener<ClusterSearchShardsResponse> listener) {
-            execute(TransportClusterSearchShardsAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ClusterSearchShardsRequestBuilder prepareSearchShards(String... indices) {
-            return new ClusterSearchShardsRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public void putRepository(PutRepositoryRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportPutRepositoryAction.TYPE, request, listener);
-        }
-
-        @Override
-        public PutRepositoryRequestBuilder preparePutRepository(String name) {
-            return new PutRepositoryRequestBuilder(this, name);
-        }
-
-        @Override
-        public ActionFuture<CreateSnapshotResponse> createSnapshot(CreateSnapshotRequest request) {
-            return execute(TransportCreateSnapshotAction.TYPE, request);
-        }
-
-        @Override
-        public void createSnapshot(CreateSnapshotRequest request, ActionListener<CreateSnapshotResponse> listener) {
-            execute(TransportCreateSnapshotAction.TYPE, request, listener);
-        }
-
-        @Override
-        public CreateSnapshotRequestBuilder prepareCreateSnapshot(String repository, String name) {
-            return new CreateSnapshotRequestBuilder(this, repository, name);
-        }
-
-        @Override
-        public CloneSnapshotRequestBuilder prepareCloneSnapshot(String repository, String source, String target) {
-            return new CloneSnapshotRequestBuilder(this, repository, source, target);
-        }
-
-        @Override
-        public void cloneSnapshot(CloneSnapshotRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportCloneSnapshotAction.TYPE, request, listener);
-        }
-
-        @Override
-        public void getSnapshots(GetSnapshotsRequest request, ActionListener<GetSnapshotsResponse> listener) {
-            execute(TransportGetSnapshotsAction.TYPE, request, listener);
-        }
-
-        @Override
-        public GetSnapshotsRequestBuilder prepareGetSnapshots(String... repositories) {
-            return new GetSnapshotsRequestBuilder(this, repositories);
-        }
-
-        @Override
-        public void deleteSnapshot(DeleteSnapshotRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportDeleteSnapshotAction.TYPE, request, listener);
-        }
-
-        @Override
-        public DeleteSnapshotRequestBuilder prepareDeleteSnapshot(String repository, String... names) {
-            return new DeleteSnapshotRequestBuilder(this, repository, names);
-        }
-
-        @Override
-        public void deleteRepository(DeleteRepositoryRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportDeleteRepositoryAction.TYPE, request, listener);
-        }
-
-        @Override
-        public DeleteRepositoryRequestBuilder prepareDeleteRepository(String name) {
-            return new DeleteRepositoryRequestBuilder(this, name);
-        }
-
-        @Override
-        public void verifyRepository(VerifyRepositoryRequest request, ActionListener<VerifyRepositoryResponse> listener) {
-            execute(VerifyRepositoryAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public VerifyRepositoryRequestBuilder prepareVerifyRepository(String name) {
-            return new VerifyRepositoryRequestBuilder(this, name);
-        }
-
-        @Override
-        public void getRepositories(GetRepositoriesRequest request, ActionListener<GetRepositoriesResponse> listener) {
-            execute(GetRepositoriesAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetRepositoriesRequestBuilder prepareGetRepositories(String... name) {
-            return new GetRepositoriesRequestBuilder(this, name);
-        }
-
-        @Override
-        public CleanupRepositoryRequestBuilder prepareCleanupRepository(String repository) {
-            return new CleanupRepositoryRequestBuilder(this, repository);
-        }
-
-        @Override
-        public void cleanupRepository(CleanupRepositoryRequest request, ActionListener<CleanupRepositoryResponse> listener) {
-            execute(TransportCleanupRepositoryAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<RestoreSnapshotResponse> restoreSnapshot(RestoreSnapshotRequest request) {
-            return execute(TransportRestoreSnapshotAction.TYPE, request);
-        }
-
-        @Override
-        public void restoreSnapshot(RestoreSnapshotRequest request, ActionListener<RestoreSnapshotResponse> listener) {
-            execute(TransportRestoreSnapshotAction.TYPE, request, listener);
-        }
-
-        @Override
-        public RestoreSnapshotRequestBuilder prepareRestoreSnapshot(String repository, String snapshot) {
-            return new RestoreSnapshotRequestBuilder(this, repository, snapshot);
-        }
-
-        @Override
-        public void snapshotsStatus(SnapshotsStatusRequest request, ActionListener<SnapshotsStatusResponse> listener) {
-            execute(TransportSnapshotsStatusAction.TYPE, request, listener);
-        }
-
-        @Override
-        public SnapshotsStatusRequestBuilder prepareSnapshotStatus(String repository) {
-            return new SnapshotsStatusRequestBuilder(this, repository);
-        }
-
-        @Override
-        public SnapshotsStatusRequestBuilder prepareSnapshotStatus() {
-            return new SnapshotsStatusRequestBuilder(this);
-        }
-
-        @Override
-        public void putPipeline(PutPipelineRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(PutPipelineTransportAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> putPipeline(PutPipelineRequest request) {
-            return execute(PutPipelineTransportAction.TYPE, request);
-        }
-
-        @Override
-        public PutPipelineRequestBuilder preparePutPipeline(String id, BytesReference source, XContentType xContentType) {
-            return new PutPipelineRequestBuilder(this, id, source, xContentType);
-        }
-
-        @Override
-        public void deletePipeline(DeletePipelineRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(DeletePipelineTransportAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> deletePipeline(DeletePipelineRequest request) {
-            return execute(DeletePipelineTransportAction.TYPE, request);
-        }
-
-        @Override
-        public DeletePipelineRequestBuilder prepareDeletePipeline(String id) {
-            return new DeletePipelineRequestBuilder(this, id);
-        }
-
-        @Override
-        public void getPipeline(GetPipelineRequest request, ActionListener<GetPipelineResponse> listener) {
-            execute(GetPipelineAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetPipelineRequestBuilder prepareGetPipeline(String... ids) {
-            return new GetPipelineRequestBuilder(this, ids);
-        }
-
-        @Override
-        public void simulatePipeline(SimulatePipelineRequest request, ActionListener<SimulatePipelineResponse> listener) {
-            execute(SimulatePipelineAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<SimulatePipelineResponse> simulatePipeline(SimulatePipelineRequest request) {
-            return execute(SimulatePipelineAction.INSTANCE, request);
-        }
-
-        @Override
-        public SimulatePipelineRequestBuilder prepareSimulatePipeline(BytesReference source, XContentType xContentType) {
-            return new SimulatePipelineRequestBuilder(this, source, xContentType);
-        }
-
-        @Override
-        public void allocationExplain(ClusterAllocationExplainRequest request, ActionListener<ClusterAllocationExplainResponse> listener) {
-            execute(TransportClusterAllocationExplainAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<ClusterAllocationExplainResponse> allocationExplain(ClusterAllocationExplainRequest request) {
-            return execute(TransportClusterAllocationExplainAction.TYPE, request);
-        }
-
-        @Override
-        public ClusterAllocationExplainRequestBuilder prepareAllocationExplain() {
-            return new ClusterAllocationExplainRequestBuilder(this);
-        }
-
-        @Override
-        public void getStoredScript(final GetStoredScriptRequest request, final ActionListener<GetStoredScriptResponse> listener) {
-            execute(GetStoredScriptAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetStoredScriptRequestBuilder prepareGetStoredScript(String id) {
-            return new GetStoredScriptRequestBuilder(this).setId(id);
-        }
-
-        @Override
-        public PutStoredScriptRequestBuilder preparePutStoredScript() {
-            return new PutStoredScriptRequestBuilder(this);
-        }
-
-        @Override
-        public void putStoredScript(final PutStoredScriptRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportPutStoredScriptAction.TYPE, request, listener);
-
-        }
-
-        @Override
-        public void deleteStoredScript(DeleteStoredScriptRequest request, ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportDeleteStoredScriptAction.TYPE, request, listener);
-        }
-
-        @Override
-        public DeleteStoredScriptRequestBuilder prepareDeleteStoredScript(String id) {
-            return new DeleteStoredScriptRequestBuilder(client).setId(id);
-        }
-    }
-
-    static class IndicesAdmin implements IndicesAdminClient {
-
-        private final ElasticsearchClient client;
-
-        IndicesAdmin(ElasticsearchClient client) {
-            this.client = client;
-        }
-
-        @Override
-        public <Request extends ActionRequest, Response extends ActionResponse> ActionFuture<Response> execute(
-            ActionType<Response> action,
-            Request request
-        ) {
-            return client.execute(action, request);
-        }
-
-        @Override
-        public <Request extends ActionRequest, Response extends ActionResponse> void execute(
-            ActionType<Response> action,
-            Request request,
-            ActionListener<Response> listener
-        ) {
-            client.execute(action, request, listener);
-        }
-
-        @Override
-        public ThreadPool threadPool() {
-            return client.threadPool();
-        }
-
-        @Override
-        public ActionFuture<IndicesAliasesResponse> aliases(final IndicesAliasesRequest request) {
-            return execute(TransportIndicesAliasesAction.TYPE, request);
-        }
-
-        @Override
-        public void aliases(final IndicesAliasesRequest request, final ActionListener<IndicesAliasesResponse> listener) {
-            execute(TransportIndicesAliasesAction.TYPE, request, listener);
-        }
-
-        @Override
-        public IndicesAliasesRequestBuilder prepareAliases() {
-            return new IndicesAliasesRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<GetAliasesResponse> getAliases(GetAliasesRequest request) {
-            return execute(GetAliasesAction.INSTANCE, request);
-        }
-
-        @Override
-        public void getAliases(GetAliasesRequest request, ActionListener<GetAliasesResponse> listener) {
-            execute(GetAliasesAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetAliasesRequestBuilder prepareGetAliases(String... aliases) {
-            return new GetAliasesRequestBuilder(this, aliases);
-        }
-
-        @Override
-        public ActionFuture<BroadcastResponse> clearCache(final ClearIndicesCacheRequest request) {
-            return execute(TransportClearIndicesCacheAction.TYPE, request);
-        }
-
-        @Override
-        public ActionFuture<GetIndexResponse> getIndex(GetIndexRequest request) {
-            return execute(GetIndexAction.INSTANCE, request);
-        }
-
-        @Override
-        public void getIndex(GetIndexRequest request, ActionListener<GetIndexResponse> listener) {
-            execute(GetIndexAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetIndexRequestBuilder prepareGetIndex() {
-            return new GetIndexRequestBuilder(this);
-        }
-
-        @Override
-        public void clearCache(final ClearIndicesCacheRequest request, final ActionListener<BroadcastResponse> listener) {
-            execute(TransportClearIndicesCacheAction.TYPE, request, listener);
-        }
-
-        @Override
-        public ClearIndicesCacheRequestBuilder prepareClearCache(String... indices) {
-            return new ClearIndicesCacheRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<CreateIndexResponse> create(final CreateIndexRequest request) {
-            return execute(TransportCreateIndexAction.TYPE, request);
-        }
-
-        @Override
-        public void create(final CreateIndexRequest request, final ActionListener<CreateIndexResponse> listener) {
-            execute(TransportCreateIndexAction.TYPE, request, listener);
-        }
-
-        @Override
-        public CreateIndexRequestBuilder prepareCreate(String index) {
-            return new CreateIndexRequestBuilder(this, index);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> delete(final DeleteIndexRequest request) {
-            return execute(TransportDeleteIndexAction.TYPE, request);
-        }
-
-        @Override
-        public void delete(final DeleteIndexRequest request, final ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportDeleteIndexAction.TYPE, request, listener);
-        }
-
-        @Override
-        public DeleteIndexRequestBuilder prepareDelete(String... indices) {
-            return new DeleteIndexRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ActionFuture<CloseIndexResponse> close(final CloseIndexRequest request) {
-            return execute(TransportCloseIndexAction.TYPE, request);
-        }
-
-        @Override
-        public void close(final CloseIndexRequest request, final ActionListener<CloseIndexResponse> listener) {
-            execute(TransportCloseIndexAction.TYPE, request, listener);
-        }
-
-        @Override
-        public CloseIndexRequestBuilder prepareClose(String... indices) {
-            return new CloseIndexRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ActionFuture<OpenIndexResponse> open(final OpenIndexRequest request) {
-            return execute(OpenIndexAction.INSTANCE, request);
-        }
-
-        @Override
-        public void open(final OpenIndexRequest request, final ActionListener<OpenIndexResponse> listener) {
-            execute(OpenIndexAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public AddIndexBlockRequestBuilder prepareAddBlock(APIBlock block, String... indices) {
-            return new AddIndexBlockRequestBuilder(this, block, indices);
-        }
-
-        @Override
-        public void addBlock(AddIndexBlockRequest request, ActionListener<AddIndexBlockResponse> listener) {
-            execute(TransportAddIndexBlockAction.TYPE, request, listener);
-        }
-
-        @Override
-        public OpenIndexRequestBuilder prepareOpen(String... indices) {
-            return new OpenIndexRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ActionFuture<BroadcastResponse> flush(final FlushRequest request) {
-            return execute(FlushAction.INSTANCE, request);
-        }
-
-        @Override
-        public void flush(final FlushRequest request, final ActionListener<BroadcastResponse> listener) {
-            execute(FlushAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public FlushRequestBuilder prepareFlush(String... indices) {
-            return new FlushRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public void getMappings(GetMappingsRequest request, ActionListener<GetMappingsResponse> listener) {
-            execute(GetMappingsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public void getFieldMappings(GetFieldMappingsRequest request, ActionListener<GetFieldMappingsResponse> listener) {
-            execute(GetFieldMappingsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetMappingsRequestBuilder prepareGetMappings(String... indices) {
-            return new GetMappingsRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ActionFuture<GetMappingsResponse> getMappings(GetMappingsRequest request) {
-            return execute(GetMappingsAction.INSTANCE, request);
-        }
-
-        @Override
-        public GetFieldMappingsRequestBuilder prepareGetFieldMappings(String... indices) {
-            return new GetFieldMappingsRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ActionFuture<GetFieldMappingsResponse> getFieldMappings(GetFieldMappingsRequest request) {
-            return execute(GetFieldMappingsAction.INSTANCE, request);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> putMapping(final PutMappingRequest request) {
-            return execute(TransportPutMappingAction.TYPE, request);
-        }
-
-        @Override
-        public void putMapping(final PutMappingRequest request, final ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportPutMappingAction.TYPE, request, listener);
-        }
-
-        @Override
-        public PutMappingRequestBuilder preparePutMapping(String... indices) {
-            return new PutMappingRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<BroadcastResponse> forceMerge(final ForceMergeRequest request) {
-            return execute(ForceMergeAction.INSTANCE, request);
-        }
-
-        @Override
-        public void forceMerge(final ForceMergeRequest request, final ActionListener<BroadcastResponse> listener) {
-            execute(ForceMergeAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ForceMergeRequestBuilder prepareForceMerge(String... indices) {
-            return new ForceMergeRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<BroadcastResponse> refresh(final RefreshRequest request) {
-            return execute(RefreshAction.INSTANCE, request);
-        }
-
-        @Override
-        public void refresh(final RefreshRequest request, final ActionListener<BroadcastResponse> listener) {
-            execute(RefreshAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public RefreshRequestBuilder prepareRefresh(String... indices) {
-            return new RefreshRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<IndicesStatsResponse> stats(final IndicesStatsRequest request) {
-            return execute(IndicesStatsAction.INSTANCE, request);
-        }
-
-        @Override
-        public void stats(final IndicesStatsRequest request, final ActionListener<IndicesStatsResponse> listener) {
-            execute(IndicesStatsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public IndicesStatsRequestBuilder prepareStats(String... indices) {
-            return new IndicesStatsRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<RecoveryResponse> recoveries(final RecoveryRequest request) {
-            return execute(RecoveryAction.INSTANCE, request);
-        }
-
-        @Override
-        public void recoveries(final RecoveryRequest request, final ActionListener<RecoveryResponse> listener) {
-            execute(RecoveryAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public RecoveryRequestBuilder prepareRecoveries(String... indices) {
-            return new RecoveryRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<IndicesSegmentResponse> segments(final IndicesSegmentsRequest request) {
-            return execute(IndicesSegmentsAction.INSTANCE, request);
-        }
-
-        @Override
-        public void segments(final IndicesSegmentsRequest request, final ActionListener<IndicesSegmentResponse> listener) {
-            execute(IndicesSegmentsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public IndicesSegmentsRequestBuilder prepareSegments(String... indices) {
-            return new IndicesSegmentsRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> updateSettings(final UpdateSettingsRequest request) {
-            return execute(TransportUpdateSettingsAction.TYPE, request);
-        }
-
-        @Override
-        public void updateSettings(final UpdateSettingsRequest request, final ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportUpdateSettingsAction.TYPE, request, listener);
-        }
-
-        @Override
-        public UpdateSettingsRequestBuilder prepareUpdateSettings(String... indices) {
-            return new UpdateSettingsRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public ActionFuture<AnalyzeAction.Response> analyze(final AnalyzeAction.Request request) {
-            return execute(AnalyzeAction.INSTANCE, request);
-        }
-
-        @Override
-        public void analyze(final AnalyzeAction.Request request, final ActionListener<AnalyzeAction.Response> listener) {
-            execute(AnalyzeAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public AnalyzeRequestBuilder prepareAnalyze(@Nullable String index, String text) {
-            return new AnalyzeRequestBuilder(this, index, text);
-        }
-
-        @Override
-        public AnalyzeRequestBuilder prepareAnalyze(String text) {
-            return new AnalyzeRequestBuilder(this, null, text);
-        }
-
-        @Override
-        public AnalyzeRequestBuilder prepareAnalyze() {
-            return new AnalyzeRequestBuilder(this);
-        }
-
-        @Override
-        public ActionFuture<AcknowledgedResponse> putTemplate(final PutIndexTemplateRequest request) {
-            return execute(TransportPutIndexTemplateAction.TYPE, request);
-        }
-
-        @Override
-        public void putTemplate(final PutIndexTemplateRequest request, final ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportPutIndexTemplateAction.TYPE, request, listener);
-        }
-
-        @Override
-        public PutIndexTemplateRequestBuilder preparePutTemplate(String name) {
-            return new PutIndexTemplateRequestBuilder(this, name);
-        }
-
-        @Override
-        public void getTemplates(final GetIndexTemplatesRequest request, final ActionListener<GetIndexTemplatesResponse> listener) {
-            execute(GetIndexTemplatesAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public GetIndexTemplatesRequestBuilder prepareGetTemplates(String... names) {
-            return new GetIndexTemplatesRequestBuilder(this, names);
-        }
-
-        @Override
-        public void deleteTemplate(final DeleteIndexTemplateRequest request, final ActionListener<AcknowledgedResponse> listener) {
-            execute(TransportDeleteIndexTemplateAction.TYPE, request, listener);
-        }
-
-        @Override
-        public DeleteIndexTemplateRequestBuilder prepareDeleteTemplate(String name) {
-            return new DeleteIndexTemplateRequestBuilder(this, name);
-        }
-
-        @Override
-        public ActionFuture<ValidateQueryResponse> validateQuery(final ValidateQueryRequest request) {
-            return execute(ValidateQueryAction.INSTANCE, request);
-        }
-
-        @Override
-        public void validateQuery(final ValidateQueryRequest request, final ActionListener<ValidateQueryResponse> listener) {
-            execute(ValidateQueryAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ValidateQueryRequestBuilder prepareValidateQuery(String... indices) {
-            return new ValidateQueryRequestBuilder(this).setIndices(indices);
-        }
-
-        @Override
-        public GetSettingsRequestBuilder prepareGetSettings(String... indices) {
-            return new GetSettingsRequestBuilder(this, indices);
-        }
-
-        @Override
-        public ResizeRequestBuilder prepareResizeIndex(String sourceIndex, String targetIndex) {
-            return new ResizeRequestBuilder(this).setSourceIndex(sourceIndex).setTargetIndex(new CreateIndexRequest(targetIndex));
-        }
-
-        @Override
-        public void resizeIndex(ResizeRequest request, ActionListener<CreateIndexResponse> listener) {
-            execute(ResizeAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public RolloverRequestBuilder prepareRolloverIndex(String alias) {
-            return new RolloverRequestBuilder(this).setRolloverTarget(alias);
-        }
-
-        @Override
-        public ActionFuture<RolloverResponse> rolloverIndex(RolloverRequest request) {
-            return execute(RolloverAction.INSTANCE, request);
-        }
-
-        @Override
-        public void rolloverIndex(RolloverRequest request, ActionListener<RolloverResponse> listener) {
-            execute(RolloverAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<GetSettingsResponse> getSettings(GetSettingsRequest request) {
-            return execute(GetSettingsAction.INSTANCE, request);
-        }
-
-        @Override
-        public void getSettings(GetSettingsRequest request, ActionListener<GetSettingsResponse> listener) {
-            execute(GetSettingsAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public void resolveIndex(ResolveIndexAction.Request request, ActionListener<ResolveIndexAction.Response> listener) {
-            execute(ResolveIndexAction.INSTANCE, request, listener);
-        }
-
-        @Override
-        public ActionFuture<ResolveIndexAction.Response> resolveIndex(ResolveIndexAction.Request request) {
-            return execute(ResolveIndexAction.INSTANCE, request);
-        }
-    }
-
-    @Override
-    public Client filterWithHeader(Map<String, String> headers) {
-        return new FilterClient(this) {
-            @Override
-            protected <Request extends ActionRequest, Response extends ActionResponse> void doExecute(
-                ActionType<Response> action,
-                Request request,
-                ActionListener<Response> listener
-            ) {
-                ThreadContext threadContext = threadPool().getThreadContext();
-                try (ThreadContext.StoredContext ctx = threadContext.stashAndMergeHeaders(headers)) {
-                    super.doExecute(action, request, listener);
-                }
-            }
-        };
-    }
-
-    /**
-     * Same as {@link PlainActionFuture} but for use with {@link RefCounted} result types. Unlike {@code PlainActionFuture} this future
-     * acquires a reference to its result. This means that the result reference must be released by a call to {@link RefCounted#decRef()}
-     * on the result before it goes out of scope.
-     * @param <R> reference counted result type
-     */
-    private static class RefCountedFuture<R extends RefCounted> extends PlainActionFuture<R> {
-
-        @Override
-        public final void onResponse(R result) {
-            result.mustIncRef();
-            if (set(result) == false) {
-                result.decRef();
-            }
-        }
-
-        private final AtomicBoolean getCalled = new AtomicBoolean(false);
-
-        @Override
-        public R get() throws InterruptedException, ExecutionException {
-            final boolean firstCall = getCalled.compareAndSet(false, true);
-            if (firstCall == false) {
-                final IllegalStateException ise = new IllegalStateException("must only call .get() once per instance to avoid leaks");
-                assert false : ise;
-                throw ise;
-            }
-            return super.get();
-        }
-    }
-}
