+++ test/framework/src/main/java/org/elasticsearch/test/client/NoOpClient.java
@@ -12,7 +12,7 @@ import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.ActionRequest;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.ActionType;
-import org.elasticsearch.client.internal.support.AbstractClient;
+import org.elasticsearch.client.internal.Client;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.threadpool.ThreadPool;
 
@@ -22,7 +22,7 @@ import org.elasticsearch.threadpool.ThreadPool;
  *
  * See also {@link NoOpNodeClient} if you need to mock a {@link org.elasticsearch.client.internal.node.NodeClient}.
  */
-public class NoOpClient extends AbstractClient {
+public class NoOpClient extends Client {
 
     public NoOpClient(ThreadPool threadPool) {
         super(Settings.EMPTY, threadPool);
